Fixed all pages
