+++ example/docker-compose.yml
@@ -8,6 +8,8 @@ services:
     volumes:
       - ./:/app
     working_dir: /app
+    ports:
+      - 1001:1001
 
   app:
     image: nginx:stable-alpine
