+++ example/src/html/local/index.html
@@ -4,12 +4,34 @@
   <head>
     <meta charset="utf-8" />
 
-    <title>TEST</title>
-    <meta name="description" content="The HTML5 Herald" />
-    <meta name="author" content="SitePoint" />
+    <title>Local implementation</title>
+    <meta
+      name="description"
+      content="Accessibility checker test page - Local implementation"
+    />
+    <meta name="author" content="Epic @ Enrise" />
+    <link rel="stylesheet" href="../../sass/index.sass" />
   </head>
 
   <body>
-    Local
+    <main>
+      <nav>
+        <a href="/">Back to main page</a>
+      </nav>
+      <p><img src="../../img/npm.svg" alt="local implementation" /></p>
+      <p>
+        This implementations runs locally, it isn't suitable to run in a
+        pipeline. <br />
+        See how we did this on the <a href="">github page</a>.
+      </p>
+      <p>
+        See the readme below:
+      </p>
+      <div
+        class="readme readme-here"
+        data-url="https://raw.githubusercontent.com/Enrise/accessibility-checker-infrastructure/master/README.md"
+      ></div>
+    </main>
+    <script src="../../js/readme.js"></script>
   </body>
 </html>
