+++ example/src/sass/index.sass
@@ -1,0 +1,7 @@
+@import 'media.sass'  
+
+@import 'main.sass'  
+@import 'stepper.sass'
+@import 'suggestion.sass'
+@import 'solutions.sass'
+@import 'readme.sass'
