+++ example/src/sass/main.sass
@@ -1,1 +1,15 @@
-@import 'stepper.sass'
+body 
+    font-family: Verdana, Geneva, Tahoma, sans-serif
+
+    & > main
+        box-sizing: border-box
+        width: 100%
+        margin: auto
+        padding: 20px
+        line-height: 1.5
+
+        @include desktop
+            width: 1000px
+
+        @include tablet
+            width: 700px
