+++ example/src/sass/readme.sass
@@ -1,0 +1,6 @@
+.readme
+    border: 1px solid #888888
+    border-radius: 5px
+    padding: 20px
+    color: #444444
+    margin-top: 20px
