+++ example/src/sass/suggestion.sass
@@ -1,0 +1,7 @@
+.suggestion
+    &__no-docker
+        display: none
+
+    &.--image
+        text-align: center
+        padding: 20px
